% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_feature_based_pathway_enrichment.R
\name{calculate_centrality}
\alias{calculate_centrality}
\title{Calculate Node Centrality in a Graph}
\usage{
calculate_centrality(graph, type = c("degree", "betweenness"))
}
\arguments{
\item{graph}{An `igraph` object representing the network.}

\item{type}{A character string specifying the type of centrality to calculate. Options are "degree" (default) or "betweenness".}
}
\value{
A named vector of centrality values for each node in the graph.
}
\description{
This function computes the centrality of nodes in a graph using either degree centrality or betweenness centrality.
}
