% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_redundance}
\alias{calculate_redundance}
\title{calculate_redundance}
\usage{
calculate_redundance(x)
}
\arguments{
\item{x}{x}
}
\description{
calculate_redundance
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
