% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metpathClass.R
\name{creatMetPathObject}
\alias{creatMetPathObject}
\title{creatMetPathObject}
\usage{
creatMetPathObject(ms1.data, sample.information, path = ".")
}
\arguments{
\item{ms1.data}{MS1 peak table name.}

\item{sample.information}{Sample information name.}

\item{path}{Work directory.}
}
\value{
A metPathClass object.
}
\description{
Creat metPathClass object.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
