% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_analysis.R
\name{doPCA}
\alias{doPCA}
\title{doPCA}
\usage{
doPCA(
  object,
  scale.method = c("no", "auto", "pareto", "center"),
  slot = c("QC", "Subject")
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{scale.method}{Scale method.}

\item{slot}{Class of data.}
}
\value{
A ggplot object.
}
\description{
PCA analysis for MetFlowData.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
