% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_analysis.R
\name{dotsne}
\alias{dotsne}
\title{dotsne}
\usage{
dotsne(
  object,
  slot = c("QC", "Subject"),
  dims = 2,
  perplexity = 30,
  colour.index = c("class", "group", "batch"),
  verbose = TRUE
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{slot}{Group of data.}

\item{dims}{The number of dimensions.}

\item{perplexity}{perplexity.}

\item{colour.index}{colour.index}

\item{verbose}{verbose}
}
\value{
A ggplot object.
}
\description{
tSNE analysis for metflowClass object.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
