% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{enrich_bar_plot}
\alias{enrich_bar_plot}
\title{enrich_bar_plot}
\usage{
enrich_bar_plot(
  object,
  x_axis = c("p_value_adjust", "p_value"),
  cutoff = 0.05,
  top = 10,
  axis.text.x.width = 15,
  axis.text.y.width = 15
)
}
\arguments{
\item{object}{enrich_result object.}

\item{x_axis}{"p_value_adjust or "p_value"}

\item{cutoff}{cutoff for p_value_adjust or p_value.}

\item{top}{top pathways to show.}

\item{axis.text.x.width}{axis.text.x.width}

\item{axis.text.y.width}{axis.text.y.width}
}
\description{
Bar plot for enrich_result
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
