% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_hmdb.R
\name{enrich_hmdb}
\alias{enrich_hmdb}
\title{enrich_kegg}
\usage{
enrich_hmdb(
  query_id,
  query_type = c("compound", "protein"),
  id_type = c("HMDB"),
  pathway_database,
  only_primary_pathway = TRUE,
  p_cutoff = 0.05,
  p_adjust_method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  method = c("hypergeometric", "fisher"),
  threads = 3
)
}
\arguments{
\item{query_id}{The vector of query IDs.}

\item{query_type}{"compound" or "gene"}

\item{id_type}{HMDB}

\item{pathway_database}{KEGG or other metabolomics pathway database.}

\item{only_primary_pathway}{only_primary_pathway}

\item{p_cutoff}{p_cutoff}

\item{p_adjust_method}{p_adjust_method}

\item{method}{Hypergeometric or fisher test.}

\item{threads}{threads}
}
\value{
The MSE analysis result.
}
\description{
Pathway enrichment for metabolomics.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
