% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_kegg.R
\name{enrich_kegg}
\alias{enrich_kegg}
\title{enrich_kegg}
\usage{
enrich_kegg(
  query_id,
  query_type = c("compound", "gene"),
  id_type = c("KEGG"),
  pathway_database,
  p_cutoff = 0.05,
  p_adjust_method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  method = c("hypergeometric", "fisher"),
  threads = 3
)
}
\arguments{
\item{query_id}{The vector of query IDs.}

\item{query_type}{"compound" or "gene"}

\item{id_type}{HMDB}

\item{pathway_database}{KEGG or other metabolomics pathway database.}

\item{p_cutoff}{p_cutoff}

\item{p_adjust_method}{p_adjust_method}

\item{method}{Hypergeometric or fisher test.}

\item{threads}{threads}
}
\value{
The MSE analysis result.
}
\description{
Pathway enrichment for SMPDB database.
}
\examples{
\dontrun{
data("kegg_hsa_pathway")

remain_idx <-
  kegg_hsa_pathway@pathway_class \%>\%
  unlist() \%>\%
  stringr::str_detect("Disease") \%>\%
  `!`() \%>\%
  which()

pathway_database <-
  filter_pathway(object = kegg_hsa_pathway,
                 remain_idx = remain_idx[1:50])

data("query_id_kegg")

object <-
  enrich_kegg(
    query_id = query_id_kegg,
    query_type = "compound",
    id_type = "KEGG",
    pathway_database = pathway_database,
    p_cutoff = 0.05,
    p_adjust_method = "BH",
    method = "hypergeometric",
    threads = 5
  )

object
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
