% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pahtway_enrichment.R
\name{enrich_pathway}
\alias{enrich_pathway}
\title{enrich_pathway}
\usage{
enrich_pathway(
  metabolite_id,
  pathway_database,
  method = c("hypergeometric", "fisher")
)
}
\arguments{
\item{metabolite_id}{The vector of gene IDs.}

\item{pathway_database}{KEGG or other metabolomics pathway database.}

\item{method}{Hypergeometric or fisher test.}
}
\value{
The MSE analysis result.
}
\description{
Pathway enrichment for metabolomics.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
