% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_result_class.R
\docType{class}
\name{enrich_result_class-class}
\alias{enrich_result_class-class}
\title{An S4 class to represent pathways}
\description{
An S4 class to represent pathways
}
\section{Slots}{

\describe{
\item{\code{pathway_database}}{pathway_database}

\item{\code{pathway_version}}{pathway_version}

\item{\code{result}}{result}
}}

