% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_result_class.R
\name{enrich_scatter_plot}
\alias{enrich_scatter_plot}
\title{enrich_bar_plot}
\usage{
enrich_scatter_plot(
  object,
  x_axis = c("mapped_percentage", "mapped_number"),
  y_axis = c("p_value_adjust", "p_value"),
  point_size = c("mapped_percentage", "all_number"),
  x_axis_cutoff = 0,
  y_axis_cutoff = 0.05,
  label = TRUE,
  label_size = 4
)
}
\arguments{
\item{object}{enrich_result_class object.}

\item{x_axis}{"mapped_number or "mapped_percentage"}

\item{y_axis}{"p_value_adjust or "p_value"}

\item{point_size}{"all_number" or "mapped_percentage"}

\item{x_axis_cutoff}{x_axis_cutoff}

\item{y_axis_cutoff}{y_axis_cutoff}

\item{label}{label}

\item{label_size}{label size}
}
\description{
Barplot for enrich_result_class
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
