% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_database-methods.R
\docType{methods}
\name{database_info,pathway_database-method}
\alias{database_info,pathway_database-method}
\alias{pathway_id,pathway_database-method}
\alias{pathway_name,pathway_database-method}
\alias{describtion,pathway_database-method}
\alias{pathway_class,pathway_database-method}
\alias{gene_list,pathway_database-method}
\alias{compound_list,pathway_database-method}
\alias{reference_list,pathway_database-method}
\alias{related_disease,pathway_database-method}
\alias{related_module,pathway_database-method}
\title{database_info method}
\usage{
\S4method{database_info}{pathway_database}(object)

\S4method{pathway_id}{pathway_database}(object)

\S4method{pathway_name}{pathway_database}(object)

\S4method{describtion}{pathway_database}(object)

\S4method{pathway_class}{pathway_database}(object)

\S4method{gene_list}{pathway_database}(object)

\S4method{compound_list}{pathway_database}(object)

\S4method{reference_list}{pathway_database}(object)

\S4method{related_disease}{pathway_database}(object)

\S4method{related_module}{pathway_database}(object)
}
\arguments{
\item{object}{(required) pathway_database class object}
}
\value{
A list (database_info)

A vector (pathway_id)

A vector (pathway_name)

A vector (describtion)

A vector (pathway_class)

A vector (gene_list)

A vector (compound_list)

A vector (reference_list)

A vector (related_disease)

A vector (related_module)
}
\description{
database_info method

pathway_id method

pathway_name method

describtion method

pathway_class method

gene_list method

compound_list method

reference_list method

related_disease method

related_module method
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}

Xiaotao Shen
\email{shenxt1990@outlook.com}

Xiaotao Shen
\email{shenxt1990@outlook.com}

Xiaotao Shen
\email{shenxt1990@outlook.com}

Xiaotao Shen
\email{shenxt1990@outlook.com}

Xiaotao Shen
\email{shenxt1990@outlook.com}

Xiaotao Shen
\email{shenxt1990@outlook.com}

Xiaotao Shen
\email{shenxt1990@outlook.com}

Xiaotao Shen
\email{shenxt1990@outlook.com}

Xiaotao Shen
\email{shenxt1990@outlook.com}
}
