% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_database_class.R
\name{filter_pathway}
\alias{filter_pathway}
\title{filter_pathway}
\usage{
filter_pathway(object, class, remain_idx)
}
\arguments{
\item{object}{pathway_database_class object.}

\item{class}{class pathway class you want to remain.}

\item{remain_idx}{Which pathways you want to remain (remain_idx).}
}
\description{
filter pathways according to pathway class
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
