% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_feature_based_pathway_enrichment.R
\name{get_hidden_metabolites}
\alias{get_hidden_metabolites}
\title{Identify Hidden Metabolites in a Metabolic Network}
\usage{
get_hidden_metabolites(
  metabolic_network,
  detected_metabolites,
  threads = 8,
  max.reaction = 3
)
}
\arguments{
\item{metabolic_network}{An `igraph` object representing the metabolic reaction network.}

\item{detected_metabolites}{A character vector of detected metabolite IDs.}

\item{threads}{Integer; number of parallel threads to use. Default is `3`.}

\item{max.reaction}{Integer; maximum number of reaction steps allowed to define a hidden metabolite. Default is `3`.}
}
\value{
A list of hidden metabolites connected to detected metabolites within the specified reaction steps.
}
\description{
This function identifies hidden metabolites that are connected to detected metabolites
within a specified number of reaction steps in a metabolic network.
}
