% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdb_database.R
\name{get_hmdb_compound}
\alias{get_hmdb_compound}
\title{get_hmdb_pathway}
\usage{
get_hmdb_compound(threads = 3)
}
\arguments{
\item{threads}{threads}
}
\description{
Get compound from HMDB (SMPDB)
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
