% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_database.R
\name{get_kegg_compound}
\alias{get_kegg_compound}
\title{get_kegg_compound}
\usage{
get_kegg_compound(local = TRUE, threads = 3)
}
\arguments{
\item{local}{Use have downloaded database or from online.}

\item{threads}{threads}
}
\description{
Get compound from KEGG
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
