% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_database.R
\name{get_kegg_pathway}
\alias{get_kegg_pathway}
\title{get_kegg_metabolite_pathway}
\usage{
get_kegg_pathway(local = TRUE, organism = "hsa", threads = 3)
}
\arguments{
\item{local}{Use have downloaded database or from online.}

\item{organism}{organism}

\item{threads}{threads}
}
\description{
Get pathway from KEGG
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
