% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_database_class.R
\name{get_pathway_class}
\alias{get_pathway_class}
\title{get_pathway_class}
\usage{
get_pathway_class(object)
}
\arguments{
\item{object}{pathway_class object.}
}
\description{
Extract the class of pathways.
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
