% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_peaks_rt}
\alias{group_peaks_rt}
\title{group peaks according to RT}
\usage{
group_peaks_rt(rt, rt.tol = 10)
}
\arguments{
\item{rt}{rt}

\item{rt.tol}{rt.tol}
}
\description{
group peaks according to RT
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
