% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{hmdb_pathway}
\alias{hmdb_pathway}
\title{HMDB Pathway Database}
\format{
A data frame with KEGG pathway data.
}
\usage{
data(hmdb_pathway)
}
\description{
A dataset containing HMDB pathways for Homo sapiens.
}
\keyword{datasets}
