% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_feature_based_pathway_enrichment.R
\name{identify_metabolic_modules}
\alias{identify_metabolic_modules}
\title{Identify Metabolic Modules}
\usage{
identify_metabolic_modules(
  sub_metabolic_network,
  detected_metabolites,
  hidden_metabolites
)
}
\arguments{
\item{sub_metabolic_network}{A metabolic network represented as a `tbl_graph` object.}

\item{detected_metabolites}{A character vector of detected metabolite IDs.}

\item{hidden_metabolites}{A character vector of hidden metabolite IDs.}
}
\value{
A list of metabolic modules, where each module is a character vector
  containing the metabolite IDs within the module.
}
\description{
This function identifies metabolic modules from a metabolic reaction network.
It applies a clustering algorithm to detect modules based on network structure.
Large modules are recursively split, and weakly connected hidden metabolites are removed.
}
