% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_compound_database}
\alias{kegg_compound_database}
\title{kegg_compound_database}
\format{
kegg_compound_database
}
\usage{
kegg_compound_database
}
\description{
kegg_compound_database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
\keyword{datasets}
