% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_hsa_pathway}
\alias{kegg_hsa_pathway}
\title{kegg_hsa_pathway}
\format{
kegg_hsa_pathway
}
\usage{
kegg_hsa_pathway
}
\description{
kegg_hsa_pathway
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
\keyword{datasets}
