% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{metpath_conflicts}
\alias{metpath_conflicts}
\title{Conflicts between the metpath and other packages}
\usage{
metpath_conflicts()
}
\value{
metpath_conflicts
}
\description{
This function lists all the conflicts between packages in the metpath
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} 
from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
metpath_conflicts()
}
