% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_metpath_logo.R
\name{metpath_logo}
\alias{metpath_logo}
\title{metpath_logo}
\usage{
metpath_logo()
}
\value{
logo
}
\description{
Get the detailed of metpath package.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
