% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{metPath_packages}
\alias{metPath_packages}
\title{List all packages in the metPath}
\usage{
metPath_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include metPath in the list?}
}
\description{
List all packages in the metPath
}
\examples{
metPath_packages()
}
