% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metpathClass.R
\docType{class}
\name{metPathClass-class}
\alias{metPathClass-class}
\title{An S4 class to represent a metPath object.}
\description{
An S4 class to represent a metPath object.
}
\section{Slots}{

\describe{
\item{\code{ms1.data}}{MS1 data.}

\item{\code{sample.info}}{Sample information.}

\item{\code{process.info}}{MS1 data.}

\item{\code{version}}{MS1 data.}
}}

