% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_class.R
\docType{class}
\name{pathway_class-class}
\alias{pathway_class-class}
\title{An S4 class to represent pathways}
\description{
An S4 class to represent pathways
}
\section{Slots}{

\describe{
\item{\code{database.info}}{Database information.}

\item{\code{spectra.info}}{Metabolites in database.}

\item{\code{spectra.data}}{MS2 spectra data.}
}}

