% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_database-class.R
\docType{class}
\name{pathway_database-class}
\alias{pathway_database-class}
\title{Pathway Database Class}
\description{
Defines the `pathway_database` class, which stores information about biological pathways, including genes, compounds, proteins, references, and related diseases.
}
\section{Slots}{

\describe{
\item{\code{database_info}}{A list containing metadata about the pathway database, such as source and version.}

\item{\code{pathway_id}}{A vector of pathway IDs.}

\item{\code{pathway_name}}{A vector of pathway names.}

\item{\code{describtion}}{A list containing descriptions of the pathways.}

\item{\code{pathway_class}}{A list categorizing pathways into different functional classes.}

\item{\code{gene_list}}{A list of genes associated with each pathway.}

\item{\code{compound_list}}{A list of compounds (metabolites) associated with each pathway.}

\item{\code{protein_list}}{A list of proteins associated with each pathway.}

\item{\code{reference_list}}{A list of references (e.g., publications) linked to each pathway.}

\item{\code{related_disease}}{A list of diseases related to each pathway.}

\item{\code{related_module}}{A list of pathway modules that are related or interconnected.}
}}

\author{
Xiaotao Shen (\email{shenxt1990@outlook.com})
}
