% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_database_class.R
\docType{class}
\name{pathway_database_class-class}
\alias{pathway_database_class-class}
\title{An S4 class to represent pathways}
\description{
An S4 class to represent pathways
}
\section{Slots}{

\describe{
\item{\code{database_info}}{database_info}

\item{\code{pathway_id}}{pathway_id}

\item{\code{pathway_name}}{pathway_name}

\item{\code{describtion}}{describtion}

\item{\code{pathway_class}}{pathway_class}

\item{\code{gene_list}}{gene_list}

\item{\code{compound_list}}{compound_list}

\item{\code{protein_list}}{protein_list}

\item{\code{reference_list}}{reference_list}

\item{\code{related_disease}}{related_disease}

\item{\code{related_module}}{related_module}
}}

