% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_feature_based_pathway_enrichment.R
\name{plot_metabolic_module_fpa}
\alias{plot_metabolic_module_fpa}
\title{Plot a Metabolic Module from FPA Results}
\usage{
plot_metabolic_module_fpa(
  fpa_result,
  feature_table_marker = feature_table_marker,
  include_feature = FALSE,
  include_hidden_metabolites = FALSE,
  add_compound_name = TRUE,
  metabolic_module_index = 1,
  layout = "fr",
  add_pathways = FALSE
)
}
\arguments{
\item{fpa_result}{List. The output of the Functional Pathway Analysis (FPA),
containing `dysregulated_metabolic_module`, `dysregulated_metabolic_network`,
and `annotation_table`.}

\item{feature_table_marker}{Data frame. The feature table containing metabolic
markers. Default is `feature_table_marker`.}

\item{include_feature}{Logical. Whether to include detected metabolic features
in the plot. Default is `FALSE`.}

\item{include_hidden_metabolites}{Logical. Whether to include hidden metabolites
in the plot. Default is `FALSE`.}

\item{add_compound_name}{Logical. Whether to add compound names as labels in
the visualization. Default is `TRUE`.}

\item{metabolic_module_index}{Numeric. The index of the metabolic module to
visualize. Default is `1`.}

\item{layout}{The layout of the network, such as `kk` or `fr`.}

\item{add_pathways}{Add pathways beside of the network or not. Default is `FALSE`.}
}
\value{
A `ggplot2` object representing the metabolic module network.
}
\description{
This function visualizes a selected metabolic module from the results of a
Functional Pathway Analysis (FPA), displaying interactions between metabolites
and features in the metabolic network.
}
\details{
The function extracts and filters a metabolic subnetwork based on the selected
module index. It then uses `ggraph` to visualize the metabolic interactions
while allowing customization, such as including metabolic features or hidden
metabolites.

The network layout follows the Fruchterman-Reingold force-directed algorithm
(`layout = "fr"`). Nodes represent metabolites and features, while edges
represent different types of metabolic interactions.
}
