% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_feature_based_pathway_enrichment.R
\name{plot_metabolic_network_fpa}
\alias{plot_metabolic_network_fpa}
\title{Plot the Dysregulated Metabolic Network from FPA Results}
\usage{
plot_metabolic_network_fpa(
  fpa_result,
  feature_table_marker,
  include_feature = FALSE,
  include_hidden_metabolites = FALSE,
  add_compound_name = TRUE
)
}
\arguments{
\item{fpa_result}{List. The output of the Functional Pathway Analysis (FPA),
containing `dysregulated_metabolic_module`, `dysregulated_metabolic_network`,
and `annotation_table`.}

\item{feature_table_marker}{Data frame. The feature table containing metabolic markers.}

\item{include_feature}{Logical. Whether to include detected metabolic features in the plot.
Default is `FALSE`.}

\item{include_hidden_metabolites}{Logical. Whether to include hidden metabolites in the plot.
Default is `FALSE`.}

\item{add_compound_name}{Logical. Whether to add compound names as labels in
the visualization. Default is `TRUE`.}
}
\value{
A `ggplot2` object representing the metabolic network.
}
\description{
This function visualizes the dysregulated metabolic network obtained from Functional Pathway Analysis (FPA),
highlighting interactions between metabolites and features in the network.
}
\details{
The function processes the metabolic network by filtering and selecting relevant nodes and edges.
It allows the inclusion of metabolic features and hidden metabolites based on user preference.
The network is visualized using the Fruchterman-Reingold force-directed algorithm (`layout = "fr"`)
to clearly illustrate metabolic interactions.
}
\examples{
\dontrun{
  # Assume `fpa_result` is a precomputed list from an FPA analysis
  plot_metabolic_network_fpa(
    fpa_result = fpa_result,
    feature_table_marker = feature_table_marker,
    include_feature = TRUE,
    include_hidden_metabolites = FALSE,
    add_compound_name = TRUE
  )
}
}
