% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{reactome_hsa_pathway}
\alias{reactome_hsa_pathway}
\title{Reactome Human Pathway Database}
\format{
A data frame with Reactome pathway data.
}
\usage{
data(reactome_hsa_pathway)
}
\description{
A dataset containing Reactome pathways for Homo sapiens.
}
\keyword{datasets}
