% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{score_peak_group}
\alias{score_peak_group}
\title{score_peak_group}
\usage{
score_peak_group(peak_group)
}
\arguments{
\item{peak_group}{peak_group}
}
\description{
score_peak_group
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
