% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_hmdb.R, R/enrich_kegg2.R
\name{set_label}
\alias{set_label}
\title{set_label}
\usage{
set_label(query_id, database, threads = parallel::detectCores() - 2)

set_label(query_id, database, threads = parallel::detectCores() - 2)
}
\arguments{
\item{query_id}{query_id}

\item{database}{database}

\item{threads}{threads}
}
\value{
The MSE analysis result.

The MSE analysis result.
}
\description{
set_label

set_label
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}

Xiaotao Shen
\email{shenxt@stanford.edu}
}
