% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pahtway_enrichment.R
\name{setlabel}
\alias{setlabel}
\title{setlabel}
\usage{
setlabel(M_ID, M)
}
\arguments{
\item{M_ID}{M_ID.}

\item{M}{M}
}
\value{
The MSE analysis result.
}
\description{
setlabel
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
