% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{wikipathway_hsa_pathway}
\alias{wikipathway_hsa_pathway}
\title{WikiPathway Human Pathway Database}
\format{
A data frame with WikiPathways pathway data.
}
\usage{
data(wikipathway_hsa_pathway)
}
\description{
A dataset containing WikiPathways pathways for Homo sapiens.
}
\keyword{datasets}
