% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_feature_based_pathway_enrichment.R
\name{calculate_redundance}
\alias{calculate_redundance}
\title{Calculate Redundancy in Metabolite Annotation}
\usage{
calculate_redundance(annotation_table)
}
\arguments{
\item{annotation_table}{A data frame containing metabolite annotations, including `Lab.ID`, `compound_class`, and `variable_id`.}
}
\value{
A numeric vector of length two:
  - The first value represents the average number of compound classes per compound (`redundancy1`).
  - The second value represents the average number of compounds assigned to each peak (`redundancy2`).
}
\description{
This function calculates redundancy metrics in a metabolite annotation table.
It evaluates the number of distinct compound classes per compound and the number
of compounds assigned to each peak.
}
