% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_result_class.R
\name{enrich_bar_plot}
\alias{enrich_bar_plot}
\title{enrich_bar_plot}
\usage{
enrich_bar_plot(
  object,
  x_axis = c("p_value_adjust", "p_value"),
  cutoff = 0.05,
  top = 10
)
}
\arguments{
\item{object}{enrich_result_class object.}

\item{x_axis}{"p_value_adjust or "p_value"}

\item{cutoff}{cutoff for p_value_adjust or p_value.}

\item{top}{top pathways to show.}
}
\description{
Barplot for enrich_result_class
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
