% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{enrich_bar_plot}
\alias{enrich_bar_plot}
\title{enrich_bar_plot}
\usage{
enrich_bar_plot(
  object,
  x_axis = c("p_value_adjust", "p_value"),
  cutoff = 0.05,
  top = 10,
  axis.text.x.width = 15,
  axis.text.y.width = 15
)
}
\arguments{
\item{object}{enrich_result object.}

\item{x_axis}{"p_value_adjust or "p_value"}

\item{cutoff}{cutoff for p_value_adjust or p_value.}

\item{top}{top pathways to show.}

\item{axis.text.x.width}{axis.text.x.width}

\item{axis.text.y.width}{axis.text.y.width}
}
\description{
Bar plot for enrich_result
}
\examples{
\dontrun{
data("kegg_hsa_pathway")
remain_idx =
  kegg_hsa_pathway@pathway_class \%>\%
  unlist() \%>\%
  stringr::str_detect("Disease") \%>\%
  `!`() \%>\%
  which()

pathway_database =
  filter_pathway(object = kegg_hsa_pathway,
                 remain_idx = remain_idx[1:50])

data("query_id_kegg")

kegg_enrichment =
  enrich_kegg(
    query_id = query_id_kegg,
    query_type = "compound",
    id_type = "KEGG",
    pathway_database = pathway_database,
    p_cutoff = 0.05,
    p_adjust_method = "BH",
    method = "hypergeometric",
    threads = 5
  )
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
