% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{enrich_network}
\alias{enrich_network}
\title{enrich_network}
\usage{
enrich_network(
  object,
  point_size = c("p_value_adjust", "p_value"),
  label = TRUE,
  label_size = 4,
  p_cutoff = 0.05,
  only_significant_pathway = TRUE,
  threads = 4
)
}
\arguments{
\item{object}{enrich_result object.}

\item{point_size}{point_size.}

\item{label}{label}

\item{label_size}{label size}

\item{p_cutoff}{p_cutoff}

\item{only_significant_pathway}{only_significant_pathway}

\item{threads}{threads}
}
\description{
Network for enriched pathways
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
