% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{enrich_scatter_plot}
\alias{enrich_scatter_plot}
\title{enrich_scatter_plot}
\usage{
enrich_scatter_plot(
  object,
  x_axis = c("mapped_percentage", "mapped_number"),
  y_axis = c("p_value_adjust", "p_value"),
  point_size = c("mapped_percentage", "all_number"),
  x_axis_cutoff = 0,
  y_axis_cutoff = 0.05,
  label = TRUE,
  label_size = 4
)
}
\arguments{
\item{object}{enrich_result object.}

\item{x_axis}{"mapped_number or "mapped_percentage"}

\item{y_axis}{"p_value_adjust or "p_value"}

\item{point_size}{"all_number" or "mapped_percentage"}

\item{x_axis_cutoff}{x_axis_cutoff}

\item{y_axis_cutoff}{y_axis_cutoff}

\item{label}{label}

\item{label_size}{label size}
}
\description{
Scatter plot for enrich_result
}
\examples{
\dontrun{
data("kegg_hsa_pathway")

remain_idx =
  kegg_hsa_pathway@pathway_class \%>\%
  unlist() \%>\%
  stringr::str_detect("Disease") \%>\%
  `!`() \%>\%
  which()

pathway_database =
  filter_pathway(object = kegg_hsa_pathway, 
                 remain_idx = remain_idx[1:50])

data("query_id_kegg")

kegg_enrichment =
  enrich_kegg(
    query_id = query_id_kegg,
    query_type = "compound",
    id_type = "KEGG",
    pathway_database = pathway_database,
    p_cutoff = 0.05,
    p_adjust_method = "BH",
    method = "hypergeometric",
    threads = 5
  )

enrich_scatter_plot(kegg_enrichment)
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
