% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_pathway_database-methods.R
\name{get_pathway_class}
\alias{get_pathway_class}
\title{Extract Pathway Class}
\usage{
get_pathway_class(object)
}
\arguments{
\item{object}{A `pathway_database` class object.}
}
\value{
A data frame with pathway classes and their counts.
}
\description{
Extracts the classification of pathways from a `pathway_database` object.
}
\author{
Xiaotao Shen (\email{shenxt1990@outlook.com})
}
