% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_feature_based_pathway_enrichment.R
\name{group_peaks_rt}
\alias{group_peaks_rt}
\title{Group Retention Time (RT) Peaks}
\usage{
group_peaks_rt(rt, rt.tol = 10)
}
\arguments{
\item{rt}{A numeric vector of retention times.}

\item{rt.tol}{Numeric; retention time tolerance for grouping. Default is `10`.}
}
\value{
A data frame with two columns:
  - `rt`: The retention times of grouped peaks.
  - `class`: The assigned group ID.
}
\description{
This function groups retention time (RT) values based on a specified tolerance,
clustering peaks that fall within the same RT range.
}
