% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{hmdb_compound_database}
\alias{hmdb_compound_database}
\title{HMDB Compound Database}
\format{
A data frame with HMDB compound details.
}
\usage{
data(hmdb_compound_database)
}
\description{
A dataset containing compound information from HMDB.
}
\keyword{datasets}
