% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdb_pathway}
\alias{hmdb_pathway}
\title{hmdb_pathway}
\format{
hmdb_pathway
}
\usage{
hmdb_pathway
}
\description{
hmdb_pathway
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
\keyword{datasets}
