% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{kegg_compound_database}
\alias{kegg_compound_database}
\title{KEGG Compound Database}
\format{
A data frame with KEGG compound details.
}
\usage{
data(kegg_compound_database)
}
\description{
A dataset containing compound information from KEGG.
}
\keyword{datasets}
