% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{kegg_hsa_pathway}
\alias{kegg_hsa_pathway}
\title{KEGG Human Pathway Database}
\format{
A data frame with KEGG pathway data.
}
\usage{
data(kegg_hsa_pathway)
}
\description{
A dataset containing KEGG pathways for Homo sapiens.
}
\keyword{datasets}
