% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_utils.R
\name{metpath_packages}
\alias{metpath_packages}
\title{List all packages in the metpath}
\usage{
metpath_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include metpath in the list?}
}
\description{
List all packages in the metpath
}
\examples{
metpath_packages()
}
