% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_database-class.R,
%   R/pathway_database-methods.R
\docType{class}
\name{pathway_database-class}
\alias{pathway_database-class}
\alias{length.pathway_database}
\alias{[.pathway_database}
\title{An S4 class to represent pathways}
\usage{
\method{length}{pathway_database}(x)

\method{[}{pathway_database}(x, i, ...)
}
\arguments{
\item{x}{x}

\item{i}{i}

\item{j}{j}

\item{drop}{drop}

\item{..}{..}
}
\value{
message

pathway_database
}
\description{
An S4 class to represent pathways
}
\section{Slots}{

\describe{
\item{\code{database_info}}{database_info}

\item{\code{pathway_id}}{pathway_id}

\item{\code{pathway_name}}{pathway_name}

\item{\code{describtion}}{describtion}

\item{\code{pathway_class}}{pathway_class}

\item{\code{gene_list}}{gene_list}

\item{\code{compound_list}}{compound_list}

\item{\code{protein_list}}{protein_list}

\item{\code{reference_list}}{reference_list}

\item{\code{related_disease}}{related_disease}

\item{\code{related_module}}{related_module}
}}

