% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{query_id_hmdb}
\alias{query_id_hmdb}
\title{HMDB Query ID List}
\format{
A character vector of HMDB compound IDs.
}
\usage{
query_id_hmdb
}
\description{
A dataset containing HMDB compound IDs for querying.
}
\author{
Xiaotao Shen (\email{shenxt1990@outlook.com})
}
\keyword{datasets}
