% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{query_id_kegg}
\alias{query_id_kegg}
\title{KEGG Query ID List}
\format{
A character vector of KEGG compound IDs.
}
\usage{
query_id_kegg
}
\description{
A dataset containing KEGG compound IDs for querying.
}
\author{
Xiaotao Shen (\email{shenxt1990@outlook.com})
}
\keyword{datasets}
