% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_feature_based_pathway_enrichment.R
\name{score_peak_group}
\alias{score_peak_group}
\title{Score a Peak Group Based on Adduct and Polarity Information}
\usage{
score_peak_group(peak_group)
}
\arguments{
\item{peak_group}{A data frame containing peak group information, including `Adduct`, `polarity`, and `isotope` columns.}
}
\value{
A numeric score representing the confidence of the peak group assignment.
}
\description{
This function assigns a score to a peak group based on the detected adduct types,
polarity, and isotopic information. Higher scores indicate a higher likelihood
of correct identification.
}
\details{
The scoring system is based on the following criteria:
- `+50` if the positive adduct is `(M+H)+`.
- `+20` if `(M+H)+` is present and is not the `[M]` isotope.
- `+20` if the polarity is positive but the adduct is not `(M+H)+`.
- `+10` if the polarity is positive, the adduct is not `(M+H)+`, and the isotope is not `[M]`.
- `+50` if the negative adduct is `(M-H)-`.
- `+20` if `(M-H)-` is present and is not the `[M]` isotope.
- `+20` if the polarity is negative but the adduct is not `(M-H)-`.
- `+10` if the polarity is negative, the adduct is not `(M-H)-`, and the isotope is not `[M]`.
}
